+++ src/Testing/MakesGraphQLRequests.php
@@ -117,11 +117,7 @@ trait MakesGraphQLRequests
      */
     protected function introspect(): TestResponse
     {
-        if (! isset($this->introspectionResult)) {
-            return $this->introspectionResult = $this->graphQL(Introspection::getIntrospectionQuery());
-        }
-
-        return $this->introspectionResult;
+        return $this->introspectionResult ??= $this->graphQL(Introspection::getIntrospectionQuery());
     }
 
     /**