+++ tests/Unit/Testing/TestingTraitDummyLumen.php
@@ -3,7 +3,8 @@
 namespace Tests\Unit\Testing;
 
 use Laravel\Lumen\Testing\Concerns\MakesHttpRequests;
-use Nuwave\Lighthouse\Testing\MakesGraphQLRequests;
+use Nuwave\Lighthouse\Testing\ClearsSchemaCache;
+use Nuwave\Lighthouse\Testing\MakesGraphQLRequestsLumen;
 use Nuwave\Lighthouse\Testing\MocksResolvers;
 use Nuwave\Lighthouse\Testing\RefreshesSchemaCache;
 use Nuwave\Lighthouse\Testing\UsesTestSchema;
@@ -12,9 +13,10 @@ use Orchestra\Testbench\TestCase;
 /**
  * Just a placeholder in order for PHPStan to be able to analyse those traits, see https://phpstan.org/blog/how-phpstan-analyses-traits.
  */
-final class TestingTraitDummy extends TestCase
+final class TestingTraitDummyLumen extends TestCase
 {
-    use MakesGraphQLRequests;
+    use ClearsSchemaCache;
+    use MakesGraphQLRequestsLumen;
     use MakesHttpRequests;
     use MocksResolvers;
     use RefreshesSchemaCache;